+++ CHANGELOG.md
@@ -76,6 +76,7 @@ Breaking changes will be documented in this changelog file for each version.
 * github-384: Improve documentation for using docker-compose
 * github-377: Load override config/env from .flowman-env.yml
 * github-344: Support .flowman-ignore file for Flowman Kernel client
+* github-385: Update Flowman tutorial
 
 ### Breaking changes